@@ -1,18 +1,18 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '27.0.3'
+    compileSdkVersion  29
+    buildToolsVersion '29.0.0'
 
     defaultConfig {
         applicationId "org.kore.kolabnotes.android"
         minSdkVersion 16
-        targetSdkVersion 26
-        versionCode 100
-        versionName "3.2.1"
+        targetSdkVersion 29
+        versionCode 101
+        versionName "4.0.0"
 
         //Running test
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     packagingOptions {
@@ -31,17 +31,17 @@ android {
 dependencies {
     //compile fileTree(dir: 'libs', include: ['*.jar'])
     //Android support libraries
-    compile 'com.android.support:support-v4:26.1.0'
-    compile 'com.android.support:appcompat-v7:26.1.0'
-    compile 'com.android.support:recyclerview-v7:26.1.0'
-    compile 'com.android.support:cardview-v7:26.1.0'
-    compile 'com.android.support:design:26.1.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
 
     // Add dependency for UI test
-    androidTestCompile 'com.android.support.test.espresso:espresso-core:3.0.1'
-    androidTestCompile 'com.android.support.test:runner:1.0.1'
-    androidTestCompile 'com.android.support:support-annotations:26.1.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.annotation:annotation:1.0.0'
 
-    compile 'com.github.konradrenner:kolabnotes-java:3.4.1'
-    compile 'jp.wasabeef:richeditor-android:1.2.2'
+    implementation 'com.github.konradrenner:kolabnotes-java:3.4.1'
+    implementation 'jp.wasabeef:richeditor-android:1.2.2'
 }
