@@ -2,8 +2,8 @@ package org.kore.kolabnotes.android;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import org.kore.kolab.notes.Attachment;
 import org.kore.kolabnotes.android.content.ActiveAccountRepository;
