@@ -1,6 +1,6 @@
 package org.kore.kolabnotes.android.adapter;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.SparseBooleanArray;
 
 import java.util.ArrayList;
