@@ -17,9 +17,9 @@ import android.print.PrintAttributes;
 import android.print.PrintDocumentAdapter;
 import android.print.PrintJob;
 import android.print.PrintManager;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.text.Html;
 import android.text.Spanned;
 import android.text.TextUtils;
@@ -78,7 +78,6 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
 
-import org.kore.kolabnotes.android.fragment.KolabNotesRichEditor;
 import yuku.ambilwarna.AmbilWarnaDialog;
 
 /**
